NumMethods

Description

numerical methods operating on T

Documentation

richardson2 :: Integer -> [T] -> T

Bender Orszag 1999 8.1.16

Given xs::[T] is a list that converges to some value, A, richardson2 n (take _N xs) evaluates the term below:

\[ \sum_{k=0}^N \frac{A_{n+k} (n+k)^N (-1)^{k+N}}{k! (N- k)!} \]

this is a bit more general than richardson. The uncertainty in the result is increased by the absolute difference between the result and the last element of xs.

something is wrong when I use this, either it's the definition, or it's being used wrong. The sums become very large and the error bound includes 0. Maybe I the A_j are terms not sums?

richardson :: (Rational -> T) -> Rational -> T

Richardson extrapolation: evaluate f at h, h/2, h/4 then infer the order and constant factor for the leading error term, and increase the h/4 term's error by this amount

numdiff2 :: (T -> T) -> T -> Rational -> T

central difference for df/dx (O(h^4) error)

numdiff :: (T -> T) -> T -> Rational -> T

central difference for df/dx (O(h^2) error)

hypergeom1f1 :: T -> T -> T -> T

wraps acb_hypgeom_1f1

secant :: (T -> T) -> Acb -> Acb -> CDouble -> T

`secant f x0 x2 tol prec` finds x such that f x == 0, between x0 and x2 Data.Number.Flint.Arb.Calc's newton or bisect may be better