NumT.IO

Description

converting T into Double and String

Documentation

captureOutputToFile :: FilePath -> IO () -> IO ()

redirect stdout to replace file contents with an action's stdout,

this shouldn't be necessary but I would have to use/find the variation on arb_printd which writes to a buffer or file

printRealparts :: (?prec :: CLong) => [T] -> IO ()

print the real part with error ie. 1.46137951490312385997144178344 +/- 5.1012e-21

printRow :: (Each s s T T, ?prec :: CLong) => s -> IO ()

printRow (a,b,c) prints value a,error a,value b,error b,value c,error c

midRad :: (?prec :: CLong) => T -> IO [Double]