Pipe

Description

Following Polyanin2002 section 3.5 or Lehktmakher1971

Given steady laminar flow in a cylindrical pipe, with a stepwise increase in wall temperature at z=0, these functions calculate the temperature field at any point in the fluid after the increase (z>0).

\[ \textbf{Pe} \left(1 - \rho^2\right) T_{,z} = T_{,\rho\rho} + \rho^{-1} T_{,\rho} + T_{,zz} \] where dimensionless variables above are defined in terms of dimensional quantities pipe radius \(R\), temperature \(T_*\), upstream wall temperature \(T_1\), downstream wall temperature \(T_2\), thermal diffusivity \(\alpha\), and maximum centerline velocity \(U_{\text{max}}\) as follows:

\[ \rho = \frac{r}{R} \] \[ z = \frac{Z}{R} \] \[ T = \frac{T_*-T_1}{T_2-T_1} \] \[ \textbf{Pe} = R U_{\text{max}} / \alpha \]

Documentation

getTemperature

Arguments

:: Integer

-> T
\(\textbf{Pe}\)

-> [(T, T)]
\(\lambda_m, A_m\)

-> [(T, T)]
\(\rho_i, z_i\)

-> IO [T]
\(T_i\)

\( T = 1 - \sum_{m=0}^{\infty} A_m \exp\left(-\frac{\lambda_m^2 z}{\textbf{Pe}} - \frac{\lambda_m \rho^2}{2}\right) {}_1F_1\left(0.5 - \frac{\lambda_m}{4} - \frac{\lambda ^3}{4 \textbf{Pe}^2}, 1, \lambda_m \rho^2\right) \) (eq. 3.5.8 and 3.5.14)

lambda

Arguments

:: T
\(\textbf{Pe}\)

-> Int
which eigenvalue m

-> CDouble
secant xtol means the result uncertainty will not be less than \(\pm \texttt{xtol}/2\).

-> T
eigenvalue \(\lambda_m\)

This starts with an approximation \( \lambda_m \approx 4m + 2.7 \) (eq. 3.5.19), and improves it with secant to satisfy

\( 0 = {}_1F_1\left(0.5 - \frac{\lambda_m}{4} - \frac{\lambda_m ^3}{4 \textbf{Pe}^2}, 1, \lambda_m \right) \) (eq 3.5.15)

_Am

Arguments

:: T
\(\textbf{Pe}\)

-> T
\(\lambda_m\)

-> Rational
h step size for numdiff2

-> T

Equation 3.5.13

\[ A_m = \frac{-2}{\lambda_m \frac{d}{d\lambda} \left[ \exp\left(-\lambda/2\right) {}_1F_1\left(0.5 - \frac{\lambda}{4} - \frac{\lambda ^3}{4 \textbf{Pe}^2}, 1, \lambda\right) \right] \Bigg|_{\lambda = \lambda_m}} \]

amDen :: T -> T -> T

5.14 denominator for _Am